backend/requirements.txt
# FastAPI & ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Caching
redis==5.0.1
celery==5.3.6
kombu==5.3.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# HTTP Client
httpx==0.26.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# NLP & Text Processing
rapidfuzz==3.6.1
tiktoken==0.5.2

# YAML Processing
pyyaml==6.0.1

# Utilities
python-dotenv==1.0.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
isort==5.13.2
mypy==1.8.0
